LayoutTests/fast/element-targeting/target-rendered-text-with-subframe-expected.txt

PASS extractedText is "zero one two three four five"
PASS successfullyParsed is true

TEST COMPLETE
This text is covered by the container
